pyproject.toml
[tool.poetry]
name = "california-pricing"
version = "0.1.0"
description = "California Pricing Price Prediction"
authors = ["GeorgeKatrilakas <georgios.katrilakas@pfizer.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
scikit-learn = "^1.5.2"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
streamlit = "^1.39.0"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
httpx = "^0.27.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
